llm_project/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0  
python-dotenv>=1.0.0       
Jinja2>=3.1.0
celery>=5.3
redis>=5.0
requests
django-q
pandas
python-dotenv